First commit
==AREAS AFFECTED

All of them! This is the first commit for the Team Site and a basic (and
incomplete) template has been pushed.

==VERSION

0.5.0

==DETAILS

So far there are 3 main pages: Homepage (index.html), Main Menu (menu.html),
On Boarding (onboarding.html). These are a work in progress but will be a part
of the final site. There is another page that was added: Login (login.php),
this may or may not be a permanent part of this site.

==REVIEWED BY

No one!